added keyboard to hang man
