+++ src/lib/lettersStore.ts
@@ -1,0 +1,4 @@
+import { writable } from 'svelte/store';
+
+export const correctLetters = writable<string[]>([]);
+export const incorrectLetters = writable<string[]>([]);
