+++ src/routes/keyboard/+page.svelte
@@ -1,0 +1,8 @@
+<script>
+import Keyboard from '$lib/Keyboard.svelte';
+
+</script>
+
+<div>
+  <Keyboard correctLetters={["d","i"]} incorrectLetters={["h","a"]}/>
+  </div>
